requirements-linux.txt
# Core telemetry
psutil==5.9.8
lxml==5.2.0
rich==13.7.1

# Embeddings / vector search
sentence-transformers==2.6.0
numpy==1.26.4
faiss-cpu==1.8.0          # choose FAISS *or* pgvector-sqlite
pgvector-sqlite==0.1.2    # (Linux only, loads vector0 extension)

# API server
fastapi==0.110.3
uvicorn==0.29.0
pydantic==1.10.14
openai==1.30.2            # or pin to the version you use

# CLI helper
typer==0.12.0

# Optional dev / test extras (keep in a separate file or pip -e .[dev])
# pytest, black, mypy
